backend/requirements.txt
annotated-types==0.6.0
asgiref==3.8.1
Django==5.0.3
django-cors-headers==4.3.1
django-ninja==1.1.0
psycopg==3.1.18
psycopg-binary==3.1.18
pydantic==2.6.4
pydantic_core==2.16.3
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
